Add newsletter builder.
The newsletter builder expects news entries to be organized in any directory
structure in the `newsletter` directory. The newsletters MUST be Markdown files,
and SHOULD contain the headers title, author, position, and date. The date
should be formatted as a string in the "yyyy-mm-dd" format.

The newsletter builder can be called with the command
`clj -A:newsletter-builder`.